+++ Changelog.md
@@ -30,6 +30,7 @@
 * Fix calculating position offset values, distance from text in different units of measurements
 * Fix parsing table style in text box
 * Fix problem with parsing absolute file path in Windows 
+* Parse `keep_next` in paragraph properties
 
 ### Refactor
 * Redone parsing images - store in structure, instead of copying file to filesystem